src/main/webapp/app/entities/mefpai/update/mefpai-update.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { FormBuilder } from '@angular/forms';
import { ActivatedRoute } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of, Subject, from } from 'rxjs';

import { MefpaiService } from '../service/mefpai.service';
import { IMefpai, Mefpai } from '../mefpai.model';

import { IUser } from 'app/entities/user/user.model';
import { UserService } from 'app/entities/user/user.service';

import { MefpaiUpdateComponent } from './mefpai-update.component';

describe('Mefpai Management Update Component', () => {
  let comp: MefpaiUpdateComponent;
  let fixture: ComponentFixture<MefpaiUpdateComponent>;
  let activatedRoute: ActivatedRoute;
  let mefpaiService: MefpaiService;
  let userService: UserService;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      declarations: [MefpaiUpdateComponent],
      providers: [
        FormBuilder,
        {
          provide: ActivatedRoute,
          useValue: {
            params: from([{}]),
          },
        },
      ],
    })
      .overrideTemplate(MefpaiUpdateComponent, '')
      .compileComponents();

    fixture = TestBed.createComponent(MefpaiUpdateComponent);
    activatedRoute = TestBed.inject(ActivatedRoute);
    mefpaiService = TestBed.inject(MefpaiService);
    userService = TestBed.inject(UserService);

    comp = fixture.componentInstance;
  });

  describe('ngOnInit', () => {
    it('Should call User query and add missing value', () => {
      const mefpai: IMefpai = { id: 456 };
      const user: IUser = { id: 63670 };
      mefpai.user = user;

      const userCollection: IUser[] = [{ id: 17661 }];
      jest.spyOn(userService, 'query').mockReturnValue(of(new HttpResponse({ body: userCollection })));
      const additionalUsers = [user];
      const expectedCollection: IUser[] = [...additionalUsers, ...userCollection];
      jest.spyOn(userService, 'addUserToCollectionIfMissing').mockReturnValue(expectedCollection);

      activatedRoute.data = of({ mefpai });
      comp.ngOnInit();

      expect(userService.query).toHaveBeenCalled();
      expect(userService.addUserToCollectionIfMissing).toHaveBeenCalledWith(userCollection, ...additionalUsers);
      expect(comp.usersSharedCollection).toEqual(expectedCollection);
    });

    it('Should update editForm', () => {
      const mefpai: IMefpai = { id: 456 };
      const user: IUser = { id: 58524 };
      mefpai.user = user;

      activatedRoute.data = of({ mefpai });
      comp.ngOnInit();

      expect(comp.editForm.value).toEqual(expect.objectContaining(mefpai));
      expect(comp.usersSharedCollection).toContain(user);
    });
  });

  describe('save', () => {
    it('Should call update service on save for existing entity', () => {
      // GIVEN
      const saveSubject = new Subject<HttpResponse<Mefpai>>();
      const mefpai = { id: 123 };
      jest.spyOn(mefpaiService, 'update').mockReturnValue(saveSubject);
      jest.spyOn(comp, 'previousState');
      activatedRoute.data = of({ mefpai });
      comp.ngOnInit();

      // WHEN
      comp.save();
      expect(comp.isSaving).toEqual(true);
      saveSubject.next(new HttpResponse({ body: mefpai }));
      saveSubject.complete();

      // THEN
      expect(comp.previousState).toHaveBeenCalled();
      expect(mefpaiService.update).toHaveBeenCalledWith(mefpai);
      expect(comp.isSaving).toEqual(false);
    });

    it('Should call create service on save for new entity', () => {
      // GIVEN
      const saveSubject = new Subject<HttpResponse<Mefpai>>();
      const mefpai = new Mefpai();
      jest.spyOn(mefpaiService, 'create').mockReturnValue(saveSubject);
      jest.spyOn(comp, 'previousState');
      activatedRoute.data = of({ mefpai });
      comp.ngOnInit();

      // WHEN
      comp.save();
      expect(comp.isSaving).toEqual(true);
      saveSubject.next(new HttpResponse({ body: mefpai }));
      saveSubject.complete();

      // THEN
      expect(mefpaiService.create).toHaveBeenCalledWith(mefpai);
      expect(comp.isSaving).toEqual(false);
      expect(comp.previousState).toHaveBeenCalled();
    });

    it('Should set isSaving to false on error', () => {
      // GIVEN
      const saveSubject = new Subject<HttpResponse<Mefpai>>();
      const mefpai = { id: 123 };
      jest.spyOn(mefpaiService, 'update').mockReturnValue(saveSubject);
      jest.spyOn(comp, 'previousState');
      activatedRoute.data = of({ mefpai });
      comp.ngOnInit();

      // WHEN
      comp.save();
      expect(comp.isSaving).toEqual(true);
      saveSubject.error('This is an error!');

      // THEN
      expect(mefpaiService.update).toHaveBeenCalledWith(mefpai);
      expect(comp.isSaving).toEqual(false);
      expect(comp.previousState).not.toHaveBeenCalled();
    });
  });

  describe('Tracking relationships identifiers', () => {
    describe('trackUserById', () => {
      it('Should return tracked User primary key', () => {
        const entity = { id: 123 };
        const trackResult = comp.trackUserById(0, entity);
        expect(trackResult).toEqual(entity.id);
      });
    });
  });
});
